Fixed an issue with shell-out command, removed chef-solo code, added timeout attribute to entry resource
In this commit, there is a few fixes. The README has been updated with current documentation adding the
timeout attribute to the resource, and clearing up the table html that was messing up presentation on the
supermarket site.

There was an error in the issues location, which has been fixed in metadata.

The cacher and default recipes had chef-solo code that has been removed.

The entry provider code around shell_out was incorrect. This was discovered by realizing that there was no
path of code that was testing whether keyscan worked. A test suite was added to .kitchen.yml to cover that
using the fixture ssh_known_hosts_test cookbook.