add a :flush action to ssh_known_hosts_entry
if this is the last action called on an ssh_known_hosts_entry
resource then the contents of the ssh_known_hosts file will
be flushed to disk.

it is the user's responsibility to ensure that they understand
their own run_list and when they can do this or not.  if you
try to flush first before calling anything else it will always
blow up (by design).  if you try to flush in the middle of
constructing your known_hosts entries then you will get
partial contents written to disk and the template resources will
flap and you will never show "0 resources updated".

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>